change dtype of embedding to long for the batcher

+++ models.py
@@ -40,7 +40,7 @@ class Model(nn.Module):
 class Baseline(nn.Module):
     def __init__(self, embeddings, device):
         super().__init__()
-        self.embeddings = nn.Embedding.from_pretrained(embeddings, freeze=True)
+        self.embeddings = nn.Embedding.from_pretrained(embeddings, freeze=True, dtype=torch.long)
 
     def forward(self, text, text_length):
         embedded = self.embeddings(text)
